Move AbilityLoadEvent to event package
